modify : user id 타입 변경

+++ src/main/java/com/risingtest/wanted/src/user/UserDao.java
@@ -63,7 +63,7 @@ public class UserDao {
         this.jdbcTemplate.update(createUserQuery, createUserParams);
 
         String lastInserIdQuery = "select last_insert_id()";
-        return this.jdbcTemplate.queryForObject(lastInserIdQuery,int.class);
+        return this.jdbcTemplate.queryForObject(lastInserIdQuery,Long.class);
     }
 
     public int checkEmail(String email){
